Extract Highlight component
